Added Schematics
Added most of the schematics to different sheets in altium project.

Still need to correct pin assignments.

Footprints need pads extended for easy soldering

There are two butterwoth schematics, use the one with 4 different op amps, need to link to the footprint of the IC, and correct the pin assignments.

Still need to add the schematics in the git and relink the libraries. RIght now they are linked to my local machine.